feat(fetch): delete --deep option
