+++ README.md
@@ -63,7 +63,6 @@ Global Flags:
       --dbtype string       Database type to store data in (sqlite3, mysql, postgres, or redis supported)
       --debug               debug mode (default: false)
       --debug-sql           SQL debug mode
-      --deep                deep mode extract cve-id from github sources
       --http-proxy string   http://proxy-url:port (default: empty)
       --log-dir string      /path/to/log
       --log-json            output log as JSON
@@ -75,14 +74,6 @@ Use "go-exploitdb fetch [command] --help" for more information about a command.
 ###### Fetch and Insert Offensive Security ExploitDB
 ```bash
 $ go-exploitdb fetch exploitdb
-```
-
-###### Deep Fetch and Insert Exploit
-- This is very time consuming
-- We will further increase the mapping rate between exploit and cveID.
-- The number of exploits that can be detected remains unchanged
-```bash
-$ go-exploitdb fetch exploitdb -deep
 ```
 
 ### Usage: Search Exploits
@@ -105,7 +96,6 @@ Global Flags:
       --dbtype string       Database type to store data in (sqlite3, mysql, postgres or redis supported)
       --debug               debug mode (default: false)
       --debug-sql           SQL debug mode
-      --deep                deep mode extract cve-id from github sources
       --http-proxy string   http://proxy-url:port (default: empty)
       --log-dir string      /path/to/log
       --log-json            output log as JSON
@@ -204,7 +194,6 @@ Global Flags:
       --dbtype string       Database type to store data in (sqlite3, mysql, postgres or redis supported)
       --debug               debug mode (default: false)
       --debug-sql           SQL debug mode
-      --deep                deep mode extract cve-id from github sources
       --http-proxy string   http://proxy-url:port (default: empty)
       --log-dir string      /path/to/log
       --log-json            output log as JSON
