+++ commands/fetch-awesomepoc.go
@@ -35,7 +35,7 @@ func fetchAwesomePoc(cmd *cobra.Command, args []string) (err error) {
 
 	log15.Info("Fetching Awesome Poc Exploit")
 	var exploits []*models.Exploit
-	if exploits, err = fetcher.FetchAwesomePoc(viper.GetBool("deep")); err != nil {
+	if exploits, err = fetcher.FetchAwesomePoc(); err != nil {
 		log15.Error("Failed to fetch AwesomePoc Exploit", "err", err)
 		return err
 	}
