+++ commands/fetch-exploitdb.go
@@ -35,7 +35,7 @@ func fetchExploitDB(cmd *cobra.Command, args []string) (err error) {
 
 	log15.Info("Fetching Offensive Security Exploit")
 	var exploits []*models.Exploit
-	if exploits, err = fetcher.FetchExploitDB(viper.GetBool("deep")); err != nil {
+	if exploits, err = fetcher.FetchExploitDB(); err != nil {
 		log15.Error("Failed to fetch Exploit", "err", err)
 		return err
 	}
