+++ commands/fetch-githubrepos.go
@@ -41,7 +41,7 @@ func fetchGitHubRepos(cmd *cobra.Command, args []string) (err error) {
 
 	log15.Info("Fetching GitHub Repos Exploit")
 	var exploits []*models.Exploit
-	if exploits, err = fetcher.FetchGitHubRepos(viper.GetBool("deep"), viper.GetInt("threshold-stars"), viper.GetInt("threshold-forks")); err != nil {
+	if exploits, err = fetcher.FetchGitHubRepos(viper.GetInt("threshold-stars"), viper.GetInt("threshold-forks")); err != nil {
 		log15.Error("Failed to fetch GitHubRepo Exploit", "err", err)
 		return err
 	}
