+++ commands/root.go
@@ -57,10 +57,6 @@ func init() {
 	_ = viper.BindPFlag("dbtype", RootCmd.PersistentFlags().Lookup("dbtype"))
 	viper.SetDefault("dbtype", "sqlite3")
 
-	RootCmd.PersistentFlags().Bool("deep", false, "deep mode extract cve-id from github sources")
-	_ = viper.BindPFlag("deep", RootCmd.PersistentFlags().Lookup("deep"))
-	viper.SetDefault("deep", false)
-
 	RootCmd.PersistentFlags().String("http-proxy", "", "http://proxy-url:port (default: empty)")
 	_ = viper.BindPFlag("http-proxy", RootCmd.PersistentFlags().Lookup("http-proxy"))
 	viper.SetDefault("http-proxy", "")
