+++ config/config.go
@@ -10,7 +10,6 @@ type CommonConfig struct {
 	Debug     bool
 	DebugSQL  bool
 	Quiet     bool
-	Deep      bool
 	DBPath    string
 	DBType    string
 	HTTPProxy string
