+++ fetcher/exploitdb.go
@@ -2,27 +2,22 @@ package fetcher
 
 import (
 	"bytes"
-	"encoding/json"
 	"encoding/xml"
 	"fmt"
-	"path/filepath"
 	"strings"
-	"time"
 
-	"github.com/cheggaaa/pb/v3"
 	"github.com/gocarina/gocsv"
 	"github.com/inconshreveable/log15"
 	"github.com/pkg/errors"
-	"github.com/vulsio/go-exploitdb/extractor"
 	"github.com/vulsio/go-exploitdb/models"
 	"github.com/vulsio/go-exploitdb/util"
 	"golang.org/x/net/html/charset"
 )
 
 // FetchExploitDB :
-func FetchExploitDB(deep bool) (exploits []*models.Exploit, err error) {
+func FetchExploitDB() (exploits []*models.Exploit, err error) {
 	var eidCvesMap map[string][]string
-	if eidCvesMap, err = FetchExploitCvesMap(deep); err != nil {
+	if eidCvesMap, err = FetchExploitCvesMap(); err != nil {
 		return nil, err
 	}
 	var exploitShellCodeMap map[string]*models.ShellCode
@@ -105,9 +100,9 @@ func FetchExploitDB(deep bool) (exploits []*models.Exploit, err error) {
 }
 
 // FetchExploitCvesMap :
-func FetchExploitCvesMap(deep bool) (eidCveMap map[string][]string, err error) {
+func FetchExploitCvesMap() (eidCveMap map[string][]string, err error) {
 	eidCveMap = map[string][]string{}
-	url := fmt.Sprintf("https://cve.mitre.org/data/downloads/allitems-cvrf.xml")
+	url := "https://cve.mitre.org/data/downloads/allitems-cvrf.xml"
 	log15.Info("Fetching", "URL", url)
 	cveXML, err := util.FetchURL(url)
 	if err != nil {
@@ -143,106 +138,7 @@ func FetchExploitCvesMap(deep bool) (eidCveMap map[string][]string, err error) {
 		}
 	}
 
-	if !deep {
-		return eidCveMap, nil
-	}
-
-	// https://github.com/offensive-security/exploitdb/search?q=CVE&unscoped_q=CVE
-	// over 500 count
-	heavyExts := []string{
-		"txt",
-		"rb",
-		"py",
-	}
-	// under 500 count
-	lightExts := []string{
-		"c",
-		"html",
-		"pl",
-		"pm",
-		"sh",
-		"md",
-		"php",
-		"cpp",
-		"java",
-		"go",
-		"cs",
-		"nse",
-		"asm",
-		"sql",
-	}
-
-	for year := 1999; year <= time.Now().Year(); year++ {
-		log15.Info("Fetching GitHub Sources", "year", year)
-		if eidCveMap, err = SearchGitHubCodeWithExt(heavyExts, year, eidCveMap); err != nil {
-			return nil, err
-		}
-	}
-	if eidCveMap, err = SearchGitHubCodeWithExt(lightExts, 0, eidCveMap); err != nil {
-		return nil, err
-	}
 	return eidCveMap, nil
-}
-
-// SearchGitHubCodeWithExt :
-func SearchGitHubCodeWithExt(exts []string, year int, exploitCveMap map[string][]string) (newExploitCveMap map[string][]string, err error) {
-	if exploitCveMap == nil {
-		exploitCveMap = map[string][]string{}
-	}
-	for _, ext := range exts {
-		page := 1
-		// max request of search code is 1000
-		maxPage := 10
-		for {
-			if maxPage < page {
-				break
-			}
-			// https://developer.github.com/v3/search/#search-code
-			var url string
-			if year == 0 {
-				url = fmt.Sprintf("https://api.github.com/search/code?q=CVE+extension:%s+repo:offensive-security/exploitdb&page=%d&per_page=100&sort=indexed&order=desc", ext, page)
-			} else {
-				spaceCode := "%20"
-				url = fmt.Sprintf("https://api.github.com/search/code?q=CVE%s%d+extension:%s+repo:offensive-security/exploitdb&page=%d&per_page=100&sort=indexed&order=desc", spaceCode, year, ext, page)
-			}
-			log15.Info("Fetching", "URL", url)
-			githubJSON, err := util.FetchURL(url)
-			if err != nil {
-				log15.Warn("Failed to fetch cve data from GitHub.", "url", url)
-				break
-			}
-			var github models.GitHubJSON
-			if err = json.Unmarshal(githubJSON, &github); err != nil {
-				return nil, err
-			}
-			bar := pb.StartNew(len(github.Items))
-			// for github rate limit
-			if 1000 < len(github.Items) {
-				log15.Warn("More than 1000 data can not be acquired due to rate limit of github")
-			}
-			for _, item := range github.Items {
-				path := "https://raw.githubusercontent.com/offensive-security/exploitdb/master/" + item.Path
-				exploitDBID := strings.TrimSuffix(filepath.Base(path), filepath.Ext(path))
-				if _, ok := exploitCveMap[exploitDBID]; !ok {
-					doc, err := util.FetchURL(path)
-					if err != nil {
-						return nil, err
-					}
-					exploitCveMap[exploitDBID] = extractor.ExtractCveID(doc)
-				}
-				bar.Increment()
-			}
-			bar.Finish()
-			if page == 1 {
-				totalPageSize := (github.TotalCount / 100) + 1
-				if totalPageSize < maxPage {
-					maxPage = totalPageSize
-				}
-			}
-			page++
-		}
-	}
-	return exploitCveMap, nil
 }
 
 // FetchExploitShellCodeMap :
