+++ fetcher/githubrepos.go
@@ -20,7 +20,7 @@ import (
 const repoURL string = "https://github.com/nomi-sec/PoC-in-GitHub.git"
 
 // FetchGitHubRepos :
-func FetchGitHubRepos(deep bool, stars, forks int) (exploits []*models.Exploit, err error) {
+func FetchGitHubRepos(stars, forks int) (exploits []*models.Exploit, err error) {
 	// clone vuln-list repository
 	dir := filepath.Join(util.CacheDir(), "pocs")
 	updatedFiles, err := git.CloneOrPull(repoURL, dir)
